[feature/fourthHomework] Changes after review;
- Create locale service for method choose language;
- Deleted interface runner service;
- Renamed mistake naming in locale config;
- Edited readMe;
