+++ fourthHomework/src/main/java/ru/javlasov/fourthHomework/service/LocaleService.java
@@ -1,0 +1,7 @@
+package ru.javlasov.fourthHomework.service;
+
+public interface LocaleService {
+
+    void chooseLanguage();
+
+}
