+++ fourthHomework/src/main/java/ru/javlasov/fourthHomework/service/TestService.java
@@ -7,6 +7,4 @@ public interface TestService {
 
     TestResult executeTestFor(Student student);
 
-    void chooseLanguage();
-
 }
